requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
openai
langchain
langchain-openai
langchain-community
langchain-core
langgraph
python-dotenv
python-multipart


